#20 io module rewritten using device manager from logic module

+++ io/src/main/kotlin/com/seraph/smarthome/io/Config.kt
@@ -27,7 +27,7 @@ enum class Parity {
 
 data class ModbusModule(
         val name: String,
-        val index: Int,
+        val index: Byte,
         val model: ModbusDeviceModel
 )
